+++ src/main/java/devkor/com/teamcback/domain/navigate/entity/Node.java
@@ -34,7 +34,7 @@ public class Node extends BaseEntity {
 
     private Double longitude;
 
-    private Integer floor;
+    private Double floor;
 
     @Column(nullable = false)
     private boolean routing;
